pyproject.toml
[project]
name = "fastapi-tdd"
version = "0.1.0"
description = ""
authors = [
    {name = "renemazuela",email = "rjmazuela@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
